+++ Sources/SwiftUIKit/Alerts/AlertContext.swift
@@ -6,58 +6,39 @@
 //  Copyright © 2020 Daniel Saidi. All rights reserved.
 //
 
-import Combine
 import SwiftUI
 
 /**
- This context can be used to present `SwiftUI` `Alert`s.
+ This context can be used to present `SwiftUI` `Alert`s. You
+ can either present plain `Alert` instances or anything that
+ implements the `AlertProvider` protocol (e.g. an enum where
+ each case returns a specific alert).
  
  To use the context, first create an observed instance, then
- bind it to your view using a regular `sheet` modifier, then
- call the context's `present` function to present an alert:
+ bind it to a view using the `alert` modifier, then call any
+ context `present` function to present an alert:
  
  ```swift
  @ObservedObject var context = AlertContext()
  
- view.alert(context: context)   // or, for more flexibility:
- view.alert(isPresented: $context.isActive, content: context.alert)
+ view.alert(context: context)
  
- context.present(Alert(title: "Hello, world!")
+ context.present(Alert(title: "Hello, world!"))
+ // ...or if an `AppAlert` enum implements `AlertProvider`:
+ context.present(AppAlert.generalWarning)
  ```
  
  The context-specific `alert` modifier is more conveinent if
  you use contexts, but the binding/content-specific one is a
  bit more flexible and can be used in other scenarios.
- 
- You can also use the `AlertProvider` protocol to present an
- alert for basically anything, e.g. an enum with cases, that
- each returns a specific alert. View the demo for more info.
  */
-public class AlertContext: ObservableObject {
+public class AlertContext: PresentationContext<Alert> {
     
-    public init() {}
-    
-    @Published public var isActive = false
-    
-    public var isActiveBinding: Binding<Bool> {
-        .init(get: { self.isActive },
-              set: { self.isActive = $0 }
-        )
+    public override func content() -> Alert {
+        contentView ?? Alert(title: Text(""))
     }
     
-    public private(set) var alertView: Alert? {
-        didSet { isActive = alertView != nil }
-    }
-    
-    public func alert() -> Alert {
-        alertView ?? Alert(title: Text(""))
-    }
-    
-    public func present(_ alert: Alert) {
-        alertView = alert
-    }
-    
-    public func present(_ alert: AlertProvider) {
-        alertView = alert.alert
+    public func present(_ provider: AlertProvider) {
+        contentView = provider.alert
     }
 }