+++ Sources/SwiftUIKit/Alerts/View+Alert.swift
@@ -14,7 +14,7 @@ public extension View {
      Present an alert from a certain alert context. The alert
      will be presented when the context is active.
      */
-    func alert<Content: View>(context: AlertContext) -> some View {
-        alert(isPresented: context.isActiveBinding, content: context.alert)
+    func alert(context: AlertContext) -> some View {
+        alert(isPresented: context.isActiveBinding, content: context.content)
     }
 }